+++ pages/login/index.js
@@ -1,13 +1,13 @@
 import Heading from "../../components/Headings/Heading";
-import HomePageNavigation from "../../components/HomepageNavigation";
 import Login from "../../components/Login";
+import Navigation from "../../components/Navigation";
 
 export default function LoginPage() {
   return (
     <>
       <Heading>Login Seite</Heading>
       <Login />
-      <HomePageNavigation />
+      <Navigation />
     </>
   );
 }
