+++ pages/teacher/index.js
@@ -1,6 +1,5 @@
 import RoomList from "../../components/RoomList";
-import Heading from "../../components/Headings/Heading";
-import HomePageNavigation from "../../components/HomepageNavigation";
+import Navigation from "../../components/Navigation";
 import { useRouter } from "next/router";
 import { useSession } from "next-auth/react";
 import { useEffect } from "react";
@@ -17,9 +16,8 @@ export default function TeacherView() {
   if (session && session.user.role === "teacher") {
     return (
       <>
-        <Heading>🏫✅ RoomCheck ✅🏫</Heading>
         <RoomList userType="teacher" />
-        <HomePageNavigation />
+        <Navigation />
       </>
     );
   }
